# Set -j option on make in librdkafka and lmdb softwares