@KevinFairise2 2023-08-28
> **❓ question** What does `-j` option do ? Configure concurrency ?

Yes `-j n` it actually tells make to create n parallel jobs to do the tasks