@KevinFairise2 2023-08-28
> Just a reminder, depending on how we calculate the value here moving to kubernetes based runners could over-parallelize this.

Yes you're right, however when migrating I configured it to use the number of CPU requested by the pod. So it should not use all the CPUs available on the host